Fix clang warnings for linux builds (#6140)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 274beb9dd42dc6c7d7dd3ff9415a41e1393f0133

+++ media/starboard/progressive/mp4_parser.cc
@@ -485,7 +485,7 @@ bool MP4Parser::ParseMP4_esds(uint64_t atom_data_size) {
   uint8_t* esds = &esds_storage[0];
   // download esds
   int bytes_read = reader_->BlockingRead(esds_offset, esds_size, esds);
-  if (bytes_read < esds_size) {
+  if (static_cast<uint64_t>(bytes_read) < esds_size) {
     LOG(WARNING) << "failed to download esds";
     return false;
   }
@@ -646,7 +646,7 @@ bool MP4Parser::ParseMP4_mvhd(uint64_t atom_data_size, uint8_t* mvhd) {
 }
 
 base::TimeDelta MP4Parser::TicksToTime(uint64_t ticks, uint32_t time_scale_hz) {
-  DCHECK_NE(time_scale_hz, 0);
+  DCHECK_NE(time_scale_hz, 0u);
 
   if (time_scale_hz == 0) {
     return base::Seconds(0);
@@ -655,7 +655,7 @@ base::TimeDelta MP4Parser::TicksToTime(uint64_t ticks, uint32_t time_scale_hz) {
 }
 
 uint64_t MP4Parser::TimeToTicks(base::TimeDelta time, uint32_t time_scale_hz) {
-  DCHECK_NE(time_scale_hz, 0);
+  DCHECK_NE(time_scale_hz, 0u);
 
   if (time_scale_hz == 0) {
     return 0;
